# [stm] Remove state handling from Futures