--- a/API/API.mli
+++ b/API/API.mli
@@ -3713,7 +3713,6 @@ sig
                    | VtProofStep of proof_step
                    | VtProofMode of string
                    | VtQuery of vernac_part_of_script * Feedback.route_id
-                   | VtBack of vernac_part_of_script * Stateid.t
                    | VtMeta
                    | VtUnknown
    and vernac_qed_type =