--- a/API/API.mli
+++ b/API/API.mli
@@ -3714,6 +3714,7 @@ sig
                    | VtProofMode of string
                    | VtQuery of vernac_part_of_script * Feedback.route_id
                    | VtBack of vernac_part_of_script * Stateid.t
+                   | VtMeta
                    | VtUnknown
    and vernac_qed_type =
      | VtKeep