--- a/library/states.ml
+++ b/library/states.ml
@@ -37,5 +37,3 @@ let with_state_protection f x =
   with reraise ->
     let reraise = CErrors.push reraise in
     (unfreeze st; iraise reraise)
-
-let with_state_protection_on_exception = Future.transactify