@ejgallego 2017-09-25
Depends on #1089 , INRIA timing job launched at:
https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/232/