@ejgallego 2017-09-25
The previous patch had some minor slowdown in some contribs, but I think we want to optimize it a bit more; testing optimized version at https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/234