@ejgallego 2017-10-05
@gares, as to get an idea of this PR, the next step in fact is making `Reach.known_state` functional, and in fact inlining `State.define` makes sense and allows for much better control of errors IMHO.

However, a critical point remains to be solved, and is the current hack done in `process_transaction` for `VtUnknown, VtNow`. This is duplicating the functionality of `Reach` and it would be great if we could implement that cleanup before landing this.

How do you think we could solve this issue? In a document model independent of scheduling that wouldn't be a problem as the scheduler would take care of this independently of whether a proof starts or not. However now, all proofs start a branch even if the end in `Define`, so indeed the case is quite tricky to handle.