@ejgallego 2017-10-05
Maybe the right fix is to get rid of `VtUnknown`. So far we have:

```
| [ "Tactic" "Notation" ltac_tactic_level_opt(n) ne_ltac_production_item_list(r) ":=" tactic(e) ] =>
  [ VtUnknown, VtNow ] ->
```

This certainly seems wrong, right @gares ?

We also have:

```
  | [ "Add" "Morphism" constr(m) ":" ident(n) ]
    (* This command may or may not open a goal *)
    => [ Vernacexpr.VtUnknown, Vernacexpr.VtNow ]
```
that is strange, what is the logic used by this command to open a proof? The logic in `add_morphism_infer` is beyond my understanding. Certainly we should fix it IMHO, even at the cost of non-backwards compatibility. @mattam82 , do you know what this command does ? The manual says it is deprecated.

And the last case is:

```
VernacInstance _ -> VtUnknown, VtNow
```

What to do with this? IMHO, it doesn't justify the mess we have now.