@gares 2017-10-09
> Maybe the right fix is to get rid of VtUnknown. So far we have:
> 
> | [ "Tactic" "Notation" ltac_tactic_level_opt(n) ne_ltac_production_item_list(r) ":=" tactic(e) ] =>
>   [ VtUnknown, VtNow ] ->
> 
> This certainly seems wrong, right @gares ?

Yes, I think it should be `VtSideff, VtNow`