@ejgallego 2017-10-09
> Yes, I think it should be VtSideff, VtNow

Ok thanks, will submit a separate PR fixing it.