@ejgallego 2017-10-12
PR fixing the ltac entry #1150