@ejgallego 2017-10-20
I guess I didn't do the proper procedure here [I thought I had marked @gares for review], but indeed, I intended for this PR not to go in until @gares had reviewed it. 

@gares , would you mind having a look to the last 3 commits in this PR and see if they are OK with you?