@maximedenes 2017-10-20
Sorry about that. If you want to be 100% sure it doesn't get merged without confirmation, you can put a "needs: review" label. But in this particular case, no metadata was showing this PR should be withheld.