@ejgallego 2017-10-20
Yup, I think I should have put a "needs: review", I think it was my fault. Anyways I think that this PR shouldn't create any problem and in fact it got merged at the proper time [early in the release], so I propose we don't revert unless @gares thinks the whole approach should be redone.