@ejgallego 2017-10-20
Yup, no rush, thanks! This PR shouldn't disturb normal Coq development in any way.