@ejgallego 2017-10-20
> I did not review the stm.ml file yet, but there is something fishy in the design. In particular Stm.States still holds cur id, but the cache seems to be moved to Vernacentries now, and the Stm has no api to invalidate it (or to force the installation of a state).
> 
> The cache holding the current system state should be in only one place IMHO.
> 
> Am I wrong? If I'm right, I'll review stm.ml once this logic is in one place.

First, indeed I think the PR should have gotten one more round of refinements, it was merged on an omission from my part.

Regarding the state, the idea is that STM becomes really functional. As such, it won't have access to the cache, which is a low-level detail of the way commands work. `interp` takes and return states, and the STM of course is free to store them and call the functions appropiadtedly. I believe the only way to have robust concurrent code is not to have imperative state [or at least pretend we don't].

So the STM will hold its table `state_id -> interp_state` as populated as it wants and call `interp` with the proper state at each moment. `State.define` and friends should go away in favor a functional traversal, and errors are handled there and thus will stop relying on the `exn_info` hack [solving quite a few bugs in the way I hope]

The real question is in what order to proceed here, manipulation of `interp_state`, `document_state` and few couple of things needs to be done and it is tricky. I have patches for both of them, and I think that making the document manipulation functional is the way to go first, [even before this PR], as the STM level cache should depend on that; however, this requires to introduce a "document owner" in the workers [in master either `coqtop` or `ide_slave` own the document] , which needs a bit of API modification.