@gares 2017-10-20
Now I see more clearly the plan, and it makes sense.
Still I think that the code of interp is wrong wrt the 1 slot cache.
Imagine the following calls, from the stm:
- Stm calls interp s1 "command that alters the state but the fails"
- The user fixes the command, and re-tries
- Stm calls interp s1 "better version of the command"

Nobody invalidated the 1 slot cache between the two calls. The STM has no API for it, and interp itself does not do it if the command fails. As a result the second command is not executed on s1 (that, even if provided, is not re-installed), but on s1' (ie s1 + the changes eventually made by the first command).