@ejgallego 2017-10-20
Your analysis is correct and that is a bug as the cache must be transactional in order to be valid.