@gares 2017-10-21
OK, then I guess I'll let you fix the thing and then re-review this PR plus the fix on vernacentries + stm